refactor addReview into api/review
